+++ bin/com/ashokit/designpattern/creation/Singtone/test.txt
