+++ src/com/ashokit/designpattern/creation/Singtone/SingletonDesignPattren.java
@@ -1,4 +1,4 @@
-package com.ashokit.designpattern;
+package com.ashokit.designpattern.creation.Singtone;
 
 import java.io.Serializable;
