+++ src/com/ashokit/designpattern/creation/Singtone/test.txt
